+++ src/geefetch/data/downloadables/collection.py
@@ -68,7 +68,7 @@ class DownloadableGEECollection(DownloadableABC):
         for key in kwargs:
             if key not in ["scale", "progress", "max_tile_size"]:
                 log.warning(f"Argument {key} is ignored.")
-        tmp_out = out.with_suffix(f".tmp.{out.suffix}")
+        tmp_out = out.with_suffix(f".tmp{out.suffix}")
         tmp_out.unlink(missing_ok=True)
         self._recursively_download(tmp_out, region, crs, bands, format)
         tmp_out.replace(out)
@@ -86,12 +86,11 @@ class DownloadableGEECollection(DownloadableABC):
         for key in kwargs:
             if key not in ["scale", "progress", "max_tile_size"]:
                 log.warning(f"Argument {key} is ignored.")
-
+        old_crs = crs
         if format == Format.GEOJSON and crs != WGS84:
             log.warning(f".geojson files must be in WGS84. Ignoring argument {crs=}.")
             crs = WGS84
-        if format == Format.PARQUET:
-            old_crs = crs
+        elif format == Format.PARQUET:
             crs = WGS84
 
         # get image download url and response
@@ -122,7 +121,7 @@ class DownloadableGEECollection(DownloadableABC):
                     self._split_then_download(
                         out,
                         region,
-                        crs,
+                        old_crs,
                         bands,
                         format,
                         _split_recursion_depth=_split_recursion_depth + 1,
@@ -187,7 +186,7 @@ class DownloadableGEECollection(DownloadableABC):
                 self._recursively_download(
                     tmp_path,
                     region,
-                    crs,
+                    WGS84,
                     bands,
                     Format.GEOJSON,
                     _split_recursion_depth,
@@ -196,6 +195,6 @@ class DownloadableGEECollection(DownloadableABC):
                 log.debug(f"Downloaded [{i + 1}/4] split for {out}.")
             gdf = merge_geojson(tmp_paths)
         if format == Format.PARQUET:
-            gdf.to_parquet(out)
+            gdf.to_crs(crs).to_parquet(out)
         else:
             gdf.to_file(out)